Selectable: Move selectable into widgets folder
